+++ database/views/intraday/intraday_by_account.sql
@@ -1,12 +1,12 @@
 DROP MATERIALIZED VIEW IF EXISTS quicksight.intraday_by_account;
 CREATE MATERIALIZED VIEW quicksight.intraday_by_account AS
 
-SELECT concat(CURRENT_DATE, ' ', coalesce(c.time_of_day, a.time_of_day))::timestamp AS trading_ts,
-       coalesce(c.trading_day, CURRENT_DATE) as trading_day,
+SELECT concat(coalesce(c.trading_day, a.trading_day), ' ', coalesce(c.time_of_day, a.time_of_day))::timestamp AS trading_ts,
+       coalesce(c.trading_day, a.trading_day) as trading_day,
        coalesce(c.time_of_day, a.time_of_day) as time_of_day,
        coalesce(c.account_id, a.account_id) as account_id,
        coalesce(c.exchange_id, a.exchange_id) as exchange_id,
-       c.updated_ts,
+       coalesce(c.updated_ts, CURRENT_TIMESTAMP::timestamp(3)) as updated_ts,
        -- current_day
        sum(c.turnover) as turnover,
        sum(c.tpl1)     as tpl1,
@@ -26,11 +26,13 @@ SELECT concat(CURRENT_DATE, ' ', coalesce(c.time_of_day, a.time_of_day))::timest
        sum(a.tpl300_avg7d_cum)   as tpl300_avg7d_cum,
        sum(a.pnl_avg7d_cum)      as pnl_avg7d_cum
 FROM quicksight._current_day c
-         FULL OUTER JOIN quicksight._avg7d a ON c.time_of_day=a.time_of_day AND c.account_id=a.account_id
-GROUP BY c.trading_day,
-         c.time_of_day, a.time_of_day,
-         c.account_id, a.account_id,
-         c.exchange_id, a.exchange_id,
-         c.updated_ts;
+FULL OUTER JOIN (
+    SELECT CURRENT_DATE as trading_day, * FROM quicksight._avg7d
+) a ON c.trading_day=a.trading_day AND c.time_of_day=a.time_of_day AND c.account_id=a.account_id
+GROUP BY coalesce(c.trading_day, a.trading_day),
+         coalesce(c.time_of_day, a.time_of_day),
+         coalesce(c.account_id,  a.account_id),
+         coalesce(c.exchange_id, a.exchange_id),
+         coalesce(c.updated_ts, CURRENT_TIMESTAMP::timestamp(3));
 
 CREATE UNIQUE INDEX intraday_by_account_idx ON quicksight.intraday_by_account (account_id, trading_ts, time_of_day);