+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 ### Added
 ### Changed
 ### Fixed
+- Fixed Duplicate issue in Materialized Views
+  - `quicksight.intraday`
+  - `quicksight.intraday_by_exchange`
+  - `quicksight.intraday_by_account`
 ### Removed
 ### -